Add detect_same_item_push to its own module
